chore(tests): move test files into __tests__ directories
src/drive/*.test.ts and src/scan/*.test.ts now live in sibling
__tests__/ dirs; import/mock specifiers adjusted one level up. No
config changes needed: vitest's default include and tsconfig's
**/*.test.ts exclude both match at any depth.
