+++ src/drive/__tests__/retry.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from 'vitest';
-import { isTransientTransportError, withTransientRetry } from './retry';
+import { isTransientTransportError, withTransientRetry } from '../retry';
 
 describe('isTransientTransportError', () => {
   it('accepts 5xx from Google, however the status is shaped', () => {
